+++ CHANGELOG.md
@@ -11,6 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - adding special parameters like "loop.first" and "loop.last" that could be used inside a Repeating block in jepy.Template
 - adding an option for validation partial to the Cached blocks in jepy.Template
 
+## [2.1.0] - 2025-06-26
+
+### Added
+
+- benchmarking with mitata
+
+### Changed
+
+- multi-line parameters will now be indented to their placeholder location
+- improved Indented block performance
+- improved Repeating block performance
+
 ## [2.0.1] - 2025-06-24
 
 ### Fixed